fixtures/requirements.txt
faker>=20.0.0
psycopg2-binary>=2.9.9
